% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{compute_vulnerability}
\alias{compute_vulnerability}
\title{Compute integrated demographic vulnerability}
\usage{
compute_vulnerability(table, column = "population_reduction")
}
\arguments{
\item{table}{Data frame returned by \code{run_grid()}.}

\item{column}{Name of the column containing percent population reduction.}
}
\value{
Mean percent population reduction, ignoring missing values.
}
\description{
Compute integrated demographic vulnerability
}
