% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{dominant_eigenvalue}
\alias{dominant_eigenvalue}
\title{Dominant eigenvalue of a projection matrix}
\usage{
dominant_eigenvalue(A)
}
\arguments{
\item{A}{Numeric square projection matrix.}
}
\value{
The real part of the eigenvalue with largest modulus.
}
\description{
Dominant eigenvalue of a projection matrix
}
