% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\name{perturbation_grid}
\alias{perturbation_grid}
\title{Perturbation grid}
\usage{
perturbation_grid(magnitudes, durations, periods)
}
\arguments{
\item{magnitudes}{Numeric vector of proportional reductions.}

\item{durations}{Integer vector of perturbation durations.}

\item{periods}{Integer vector of perturbation periods.}
}
\value{
An object of class \code{demovuln_grid}.
}
\description{
Perturbation grid
}
