% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{population_reduction}
\alias{population_reduction}
\title{Compute percent population reduction}
\usage{
population_reduction(final_population, baseline_final_population)
}
\arguments{
\item{final_population}{Final population size under perturbed dynamics.}

\item{baseline_final_population}{Final population size under unperturbed
baseline dynamics.}
}
\value{
Percent population reduction relative to the baseline.
}
\description{
Compute percent population reduction
}
